feat: add final id for VIP 186 and description
