+++ vips/vip-186/vip-186-testnet.ts
@@ -1,17 +1,17 @@
 import { ProposalType } from "../../src/types";
 import { makeProposal } from "../../src/utils";
 
-const COMPTROLLER_BEACON = "0x38B4Efab9ea1bAcD19dC81f19c4D1C2F9DeAe1B2";
-const NEW_COMPTROLLER_IMPLEMENTATION = "0x69Ca940186C29b6a9D64e1Be1C59fb7A466354E2";
-const ACM = "0x4788629ABc6cFCA10F9f969efdEAa1cF70c23555";
-const POOL_STABLECOIN = "0x94c1495cD4c557f1560Cbd68EAB0d197e6291571";
-const POOL_DEFI = "0x3344417c9360b963ca93A4e8305361AEde340Ab9";
-const POOL_GAMEFI = "0x1b43ea8622e76627B81665B1eCeBB4867566B963";
-const POOL_LIQUID_STAKED_BNB = "0xd933909A4a2b7A4638903028f44D1d38ce27c352";
-const POOL_TRON = "0x23b4404E4E5eC5FF5a6FFb70B7d14E3FabF237B0";
-const CRITICAL_TIMELOCK = "0x213c446ec11e45b15a6E29C1C1b402B8897f606d";
-const NORMAL_TIMELOCK = "0x939bD8d64c0A9583A7Dcea9933f7b21697ab6396";
-const FAST_TRACK_TIMELOCK = "0x555ba73dB1b006F3f2C7dB7126d6e4343aDBce02";
+const COMPTROLLER_BEACON = "0xdddd7725c073105fb2abfcbdec16708fc4c24b74";
+const NEW_COMPTROLLER_IMPLEMENTATION = "0x11a92852fA7D70C220Dada69969b2f1C4e18e663";
+const ACM = "0x45f8a08F534f34A97187626E05d4b6648Eeaa9AA";
+const POOL_STABLECOIN = "0x10b57706AD2345e590c2eA4DC02faef0d9f5b08B";
+const POOL_DEFI = "0x23a73971A6B9f6580c048B9CB188869B2A2aA2aD";
+const POOL_GAMEFI = "0x1F4f0989C51f12DAcacD4025018176711f3Bf289";
+const POOL_LIQUID_STAKED_BNB = "0x596B11acAACF03217287939f88d63b51d3771704";
+const POOL_TRON = "0x11537D023f489E4EF0C7157cc729C7B69CbE0c97";
+const FAST_TRACK_TIMELOCK = "0x3CFf21b7AF8390fE68799D58727d3b4C25a83cb6";
+const CRITICAL_TIMELOCK = "0x23B893a7C45a5Eb8c8C062b9F32d0D2e43eD286D";
+const NORMAL_TIMELOCK = "0xce10739590001705F7FF231611ba4A48B2820327";
 
 interface GrantAccess {
   target: string;
@@ -32,12 +32,12 @@ const grantAccessControl = (contract: string) => {
   return accessProposals;
 };
 
-export const vip185 = () => {
+export const vip186Testnet = () => {
   const meta = {
     version: "v2",
-    title: "VIP-185 Add forced liquidations feature into the Isolated Pools",
+    title: "VIP comptroller beacon upgrade",
     description: `
-        Upgrades the implementation contract for comptroller beacon for isolated pools
+        Upgardes the implementation contract for comptroller beacon for isolated pools
         Gives call permissions to all three timelocks for setting forced liquidation`,
 
     forDescription: "I agree that Venus Protocol should proceed with upgrading comptroller beacon",
